--- test/cbmc/proofs/DNS/NBNSHandlePacket/Makefile.json
@@ -1,23 +1,0 @@
-{
-  "ENTRY": "NBNSHandlePacket",
-  "USE_CACHE":1,
-  "CBMCFLAGS":
-  [
-  	"--unwind 1",
-    "--unwindset DNS_TreatNBNS.0:16",
-  	"--nondet-static"
-  ],
-  "OBJS":
-  [
-    "$(ENTRY)_harness.goto",
-    "$(FREERTOS_PLUS_TCP)/test/cbmc/stubs/cbmc.goto",
-    "$(FREERTOS_PLUS_TCP)/source/FreeRTOS_DNS.goto",
-    "$(FREERTOS_PLUS_TCP)/source/FreeRTOS_IP_Utils.goto",
-    "$(FREERTOS_PLUS_TCP)/source/FreeRTOS_DNS_Parser.goto"
-  ],
-  "DEF":
-  [
-    "ipconfigUSE_DNS_CACHE={USE_CACHE}",
-    "ipconfigUSE_NBNS=1"
-  ]
-}
